// src/app/models/maintenance/device-serial-mapping/deviceSerialMap.ts
export class DeviceSerial {
    MANUFACTURER_ID: string;
    RANGE_START: string;
    RANGE_END: string;
    PROGRAM_NAME: string;
    PROGRAM_INDICATOR: string;
    LOOKUP_TYPE: string;
    LOOKUP_REF: string;
    ACTIVE: string;
    SEQID: string;
    CLIENTID: string;
    SITEID: string;
    ADDTS: string;
    ADDWHO: string;
    EDITTS: string;
    EDITWHO: string;
}